requirements.txt
python-telegram-bot[job-queue]==20.8
python-dotenv==1.0.0 